Replace FileLoader Table with react-virtualized Table
