+++ app/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
+import 'react-virtualized/styles.css';
 import Root from './containers/Root';
 import { configureStore, history } from './store/configureStore';
 import { initializeFirebase } from './domain/FirebaseAuthentication';
